pull out valid_token? to SkateWeb.ChannelAuth
